# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Beckbest Bridal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers/register:
    post:
      operationId: registerCustomer
      tags: [customers]
      summary: Register a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegistration"
      responses:
        "201":
          description: Customer registered, verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          description: Invalid input
        "409":
          description: Email already exists
  /customers/verify-email:
    post:
      operationId: verifyCustomerEmail
      tags: [customers]
      summary: Verify email with 6-digit code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified, returns auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired code
  /customers/resend-verification:
    post:
      operationId: resendVerification
      tags: [customers]
      summary: Resend email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationInput"
      responses:
        "200":
          description: Code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /customers/forgot-password:
    post:
      operationId: forgotPassword
      tags: [customers]
      summary: Request password reset code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Reset code sent if account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /customers/verify-reset-code:
    post:
      operationId: verifyResetCode
      tags: [customers]
      summary: Verify 6-digit reset code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyResetCodeInput"
      responses:
        "200":
          description: Code verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          description: Invalid or expired code
  /customers/reset-password:
    post:
      operationId: resetPassword
      tags: [customers]
      summary: Reset password with verified code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          description: Invalid request
  /customers/login:
    post:
      operationId: loginCustomer
      tags: [customers]
      summary: Login as customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
        "403":
          description: Email not verified
...
  /customers/me:
    get:
      operationId: getMe
      tags: [customers]
      summary: Get current customer profile
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    Customer:
      type: object
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: ["string", "null"] }
        isVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, email, firstName, lastName, isVerified, createdAt]
    CustomerRegistration:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        phone: { type: string }
      required: [email, password, firstName, lastName]
    LoginInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]
    VerifyEmailInput:
      type: object
      properties:
        email: { type: string, format: email }
        code: { type: string, minLength: 6, maxLength: 6 }
      required: [email, code]
    ResendVerificationInput:
      type: object
      properties:
        email: { type: string, format: email }
      required: [email]
    EmailInput:
      type: object
      properties:
        email: { type: string, format: email }
      required: [email]
    VerifyResetCodeInput:
      type: object
      properties:
        email: { type: string, format: email }
        code: { type: string, minLength: 6, maxLength: 6 }
      required: [email, code]
    ResetPasswordInput:
      type: object
      properties:
        email: { type: string, format: email }
        code: { type: string, minLength: 6, maxLength: 6 }
        newPassword: { type: string, minLength: 8 }
      required: [email, code, newPassword]
    AuthResponse:
      type: object
      properties:
        customer: { $ref: "#/components/schemas/Customer" }
        token: { type: string }
      required: [customer, token]
...